// src/api/audit/models/AuditLogResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type AuditLogResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string | undefined
   */
  serviceName?: string;
  /**
   * @type string | undefined
   */
  method?: string;
  /**
   * @type string | undefined
   */
  url?: string;
  /**
   * @type object
   */
  body?: {
    [key: string]:
      | string
      | number
      | boolean
      | null
      | (
          | string
          | number
          | boolean
          | null
          | {
              [key: string]: string | number | boolean | null;
            }
        )[]
      | {
          [key: string]:
            | string
            | number
            | boolean
            | null
            | (string | number | boolean | null)[];
        };
  } | null;
  /**
   * @type object
   */
  query?: {
    [key: string]:
      | string
      | number
      | boolean
      | null
      | (
          | string
          | number
          | boolean
          | null
          | {
              [key: string]: string | number | boolean | null;
            }
        )[]
      | {
          [key: string]:
            | string
            | number
            | boolean
            | null
            | (string | number | boolean | null)[];
        };
  } | null;
  /**
   * @type object
   */
  headers?: {
    [key: string]:
      | string
      | number
      | boolean
      | null
      | (
          | string
          | number
          | boolean
          | null
          | {
              [key: string]: string | number | boolean | null;
            }
        )[]
      | {
          [key: string]:
            | string
            | number
            | boolean
            | null
            | (string | number | boolean | null)[];
        };
  } | null;
  /**
   * @type object
   */
  userId?: object | null;
  /**
   * @type object
   */
  userName?: object | null;
  /**
   * @type object
   */
  userLogin?: object | null;
  /**
   * @type string | undefined
   */
  ip?: string;
  /**
   * @type number | undefined
   */
  statusCode?: number;
  /**
   * @type number | undefined
   */
  durationMs?: number;
  /**
   * @type object
   */
  errorMessage?: object | null;
  /**
   * @type string, date-time
   */
  timestamp: Date;
  /**
   * @type string, date-time
   */
  receivedAt: Date;
  /**
   * @type string, date-time
   */
  createdAt: Date;
};