// src/api/audit/models/PaginatedAuditLogResponseDtoDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AuditLogResponseDto } from './AuditLogResponseDto';

export type PaginatedAuditLogResponseDtoDto = {
  /**
   * @type number
   */
  total: number;
  /**
   * @type number
   */
  page: number;
  /**
   * @type number
   */
  limit: number;
  /**
   * @type number
   */
  totalPages: number;
  /**
   * @description Lista de itens paginados do tipo AuditLogResponseDto.
   * @type array
   */
  data: AuditLogResponseDto[];
};